Switching more things to from long/double->int/float to save space. Still trying to figure out livelock issue...
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/markov-branch@1846 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ src/frontend/edu/brown/markov/Edge.java
@@ -32,14 +32,14 @@ public class Edge extends AbstractEdge implements Comparable<Edge> {
     /**
      * This is the total number of times that we have traversed over this edge
      */
-    public long hits;
+    public int hits;
 
     /**
      * This is the temporary number of times that we have traversed over this edge in the current "period" of the
      * MarkovGraph. This will eventually get folded into the global hits count, but we need to keep it separate so that
      * we can determine whether the current workload is deviating from the training set
      */
-    private transient long instancehits = 0;
+    private transient int instancehits = 0;
 
     /**
      * Constructor
